Explicitly targeting .net core app 2.0, conditionally compiling; fixes #2680
