+++ src/AutoMapper/Execution/PropertyEmitter.cs
@@ -1,3 +1,4 @@
+#if DYNAMIC_METHODS
 namespace AutoMapper.Execution
 {
     using System;
@@ -65,3 +66,4 @@ namespace AutoMapper.Execution
             : _setterBuilder;
     }
 }
+#endif
