+++ src/AutoMapper/Execution/ProxyGenerator.cs
@@ -1,3 +1,4 @@
+#if DYNAMIC_METHODS
 namespace AutoMapper.Execution
 {
     using System;
@@ -116,8 +117,7 @@ namespace AutoMapper.Execution
             var fieldBuilders = new Dictionary<string, PropertyEmitter>();
             foreach(var property in propertiesToImplement)
             {
-                PropertyEmitter propertyEmitter;
-                if(fieldBuilders.TryGetValue(property.Name, out propertyEmitter))
+                if(fieldBuilders.TryGetValue(property.Name, out var propertyEmitter))
                 {
                     if((propertyEmitter.PropertyType != property.Type) &&
                         ((property.CanWrite) || (!property.Type.IsAssignableFrom(propertyEmitter.PropertyType))))
@@ -130,8 +130,7 @@ namespace AutoMapper.Execution
                 else
                 {
                     fieldBuilders.Add(property.Name,
-                        propertyEmitter =
-                            new PropertyEmitter(typeBuilder, property, propertyChangedField));
+                        new PropertyEmitter(typeBuilder, property, propertyChangedField));
                 }
             }
             return typeBuilder.CreateType();
@@ -161,83 +160,5 @@ namespace AutoMapper.Execution
             return bytes;
         }
     }
-
-    public struct TypeDescription : IEquatable<TypeDescription>
-    {
-        public TypeDescription(Type type) : this(type, PropertyDescription.Empty)
-        {
-        }
-
-        public TypeDescription(Type type, IEnumerable<PropertyDescription> additionalProperties)
-        {
-            Type = type ?? throw new ArgumentNullException(nameof(type));
-            if(additionalProperties == null)
-            {
-                throw new ArgumentNullException(nameof(additionalProperties));
-            }
-            AdditionalProperties = additionalProperties.OrderBy(p => p.Name).ToArray();
-        }
-
-        public Type Type { get; }
-
-        public PropertyDescription[] AdditionalProperties { get; }
-
-        public override int GetHashCode()
-        {
-            var hashCode = Type.GetHashCode();
-            foreach(var property in AdditionalProperties)
-            {
-                hashCode = HashCodeCombiner.CombineCodes(hashCode, property.GetHashCode());
-            }
-            return hashCode;
-        }
-
-        public override bool Equals(object other) => other is TypeDescription && Equals((TypeDescription)other);
-
-        public bool Equals(TypeDescription other) => Type == other.Type && AdditionalProperties.SequenceEqual(other.AdditionalProperties);
-
-        public static bool operator ==(TypeDescription left, TypeDescription right) => left.Equals(right);
-
-        public static bool operator !=(TypeDescription left, TypeDescription right) => !left.Equals(right);
-    }
-
-    [DebuggerDisplay("{Name}-{Type.Name}")]
-    public struct PropertyDescription : IEquatable<PropertyDescription>
-    {
-        internal static PropertyDescription[] Empty = new PropertyDescription[0];
-
-        public PropertyDescription(string name, Type type, bool canWrite = true)
-        {
-            Name = name;
-            Type = type;
-            CanWrite = canWrite;
-        }
-
-        public PropertyDescription(PropertyInfo property)
-        {
-            Name = property.Name;
-            Type = property.PropertyType;
-            CanWrite = property.CanWrite;
-        }
-
-        public string Name { get; }
-
-        public Type Type { get; }
-
-        public bool CanWrite { get; }
-
-        public override int GetHashCode()
-        {
-            var code = HashCodeCombiner.Combine(Name, Type);
-            return HashCodeCombiner.CombineCodes(code, CanWrite.GetHashCode());
-        }
-
-        public override bool Equals(object other) => other is PropertyDescription && Equals((PropertyDescription)other);
-
-        public bool Equals(PropertyDescription other) => Name == other.Name && Type == other.Type && CanWrite == other.CanWrite;
-
-        public static bool operator ==(PropertyDescription left, PropertyDescription right) => left.Equals(right);
-
-        public static bool operator !=(PropertyDescription left, PropertyDescription right) => !left.Equals(right);
-    }
 }
+#endif
