+++ src/AutoMapper/LockingConcurrentDictionary.cs
@@ -31,7 +31,7 @@ namespace AutoMapper
                 value = lazy.Value;
                 return true;
             }
-            value = default(TValue);
+            value = default;
             return false;
         }
