+++ src/AutoMapper/PropertyMap.cs
@@ -17,7 +17,7 @@ namespace AutoMapper
         private readonly List<MemberInfo> _memberChain = new List<MemberInfo>();
         private readonly List<ValueTransformerConfiguration> _valueTransformerConfigs = new List<ValueTransformerConfiguration>();
 
-        internal static PropertyMap Default { get; } = new PropertyMap(default(MemberInfo), default(TypeMap));
+        internal static PropertyMap Default { get; } = new PropertyMap(default(MemberInfo), default);
         
         public PropertyMap(PathMap pathMap)
         {
