+++ src/UnitTests/Internal/CreateProxyThreading.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿#if DYNAMIC_METHODS
+using System;
 using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper.Execution;
@@ -35,3 +36,4 @@ namespace AutoMapper.UnitTests
 
     }
 }
+#endif
